status code

+++ Orchestration/Controllers/OrchestratorController.cs
@@ -21,25 +21,20 @@ namespace Orchestration.Controllers
             // entry in identity db
             var identityClient = _httpClientFactory?.CreateClient("identity");
             var identityData = new StringContent( JsonSerializer.Serialize(name), Encoding.UTF8, Application.Json);
-            var identityResponse = await identityClient?.PostAsync("https://localhost:7249/users", identitydata);
+            var identityResponse = await identityClient?.PostAsync("https://localhost:7249/users", identityData);
 
-            var identityId=await identityResponse.Content.ReadAsStringAsync();
+            var identityId=await identityResponse.Content.ReadAsStringAsync(); 
 
-            try {
+            // entry in license db
+            var licenseClient = _httpClientFactory?.CreateClient("license");
+            var licenseData = new StringContent(JsonSerializer.Serialize(license), Encoding.UTF8, Application.Json);
+            var licenseResponse = await licenseClient?.PostAsync("https://localhost:7194/license", licenseData);
 
-                // entry in license db
-                var licenseClient = _httpClientFactory?.CreateClient("license");
-                var licenseData = new StringContent(JsonSerializer.Serialize(license), Encoding.UTF8, Application.Json);
-                var licenseResponse = await licenseClient?.PostAsync("https://localhost:7194/license", licenseData);
-
+            // compensating transaction if license db entry fails
+            if (licenseResponse.StatusCode== System.Net.HttpStatusCode.InternalServerError) {            
+                    var deleteResponse = await identityClient?.DeleteAsync("https://localhost:7249/users/" + identityId);
+                    return BadRequest();
             }
-            catch (Exception ex) {
-
-                // compensating transaction if license entry fails
-                var deleteResponse = await identityClient?.DeleteAsync("https://localhost:7249/users/" + identityId);
-                return BadRequest();
-            }
-
 
             return Ok();
         }
